LPD-34724 move CTCollectionPreviewThreadLocal to portal-kernel
