+++ portal-kernel/src/com/liferay/portal/kernel/change/tracking/CTCollectionPreviewThreadLocal.java
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.change.tracking.internal;
+package com.liferay.portal.kernel.change.tracking;
 
 import com.liferay.petra.lang.CentralizedThreadLocal;
-import com.liferay.portal.kernel.change.tracking.CTCollectionThreadLocal;
 
 /**
  * @author David Truong
